add version 2
Jobs have been deleted using the button under the job. Checked for the number of requests so that more than available cannot be done. It has been made so that after a restart, the background job search starts automatically. Added functionality for adding unique vacancies that are only available in the bot. Added unique vacancies in the background search too